+++ experimental/packages/exporter-trace-otlp-http/src/platform/browser/OTLPTraceExporter.ts
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-import { OTLPExporterConfigBase } from '../../types';
-import { OTLPExporterBrowserBase } from './OTLPExporterBrowserBase';
+import { appendResourcePathToUrlIfNotPresent, OTLPExporterBrowserBase } from '@opentelemetry/otlp-exporter-base';
 import { ReadableSpan, SpanExporter } from '@opentelemetry/sdk-trace-base';
 import { toOTLPExportTraceServiceRequest } from '../../transform';
 import * as otlpTypes from '../../types';
 import { getEnv, baggageUtils } from '@opentelemetry/core';
-import { appendResourcePathToUrlIfNotPresent } from '../../util';
+import { OTLPExporterConfigBase } from '@opentelemetry/otlp-exporter-base';
 
 const DEFAULT_COLLECTOR_RESOURCE_PATH = '/v1/traces';
 const DEFAULT_COLLECTOR_URL=`http://localhost:4318${DEFAULT_COLLECTOR_RESOURCE_PATH}`;